Update PokemonController.php
Made changes to the Pokemon Controller to protect against the application breaking due to 'artisan php serve' not correctly setting up the environment variable. This may be due to something I am missing but I believe it is a bug the version of laravel that I am using. The inclusion of the "safeURL' helper and updating of the code to make use of it should ensure that it'll always try to use the envrionment 'APP_URL', but will default to localhost:8000 if that errors out.