Added Captures
Implemented captures into the system. Used https://tutsforweb.com/laravel-passport-create-rest-api-with-authentication/ as a base to build onto. Users can now define a pokemon as captured. They can retrieve a list of all the pokemon they have marked as captured and delete captures, marking the pokemon as no longer captured.  Continued improvements in the formatting of the response are ongoing. WIP